fix(blender): keep modern 4.2 import pipeline
