+++ FortnitePorting.Plugins/Blender/fortnite_porting/processing/legacy/context.py
@@ -1,25 +1,32 @@
-from ..enums import EExportType as CurrentExportType
-from ..enums import EPrimitiveExportType as CurrentPrimitiveExportType
-from .enums import EExportType as LegacyExportType
-from .enums import EPrimitiveExportType as LegacyPrimitiveExportType
-from .import_context import LegacyMaterialImportContext
+from ..context import (
+    AnimImportContext,
+    BaseImportContext,
+    FontImportContext,
+    MeshImportContext,
+    PoseImportContext,
+    SoundImportContext,
+    TastyImportContext,
+    TextureImportContext,
+)
+from .data import ensure_legacy_blend_data
+from .material_context import LegacyMaterialImportContext
 
 
-class LegacyImportContext(LegacyMaterialImportContext):
-    """Imports on Blender 4.2-4.5 using the compatible V3 asset library."""
+class LegacyImportContext(
+    BaseImportContext,
+    MeshImportContext,
+    AnimImportContext,
+    TextureImportContext,
+    SoundImportContext,
+    FontImportContext,
+    PoseImportContext,
+    TastyImportContext,
+    LegacyMaterialImportContext,
+):
+    """Uses the modern import pipeline with Blender 4.2-compatible materials."""
 
-    def run(self, data):
-        legacy_data = dict(data)
+    def __init__(self, meta_data):
+        BaseImportContext.__init__(self, meta_data)
 
-        try:
-            export_name = CurrentExportType(data.get("Type")).name
-            legacy_data["Type"] = LegacyExportType[export_name].value
-
-            primitive_name = CurrentPrimitiveExportType(data.get("PrimitiveType")).name
-            legacy_data["PrimitiveType"] = LegacyPrimitiveExportType[primitive_name].value
-        except (KeyError, ValueError) as exc:
-            raise ValueError(
-                "This export type is not supported by the Blender 4.2-compatible import pipeline."
-            ) from exc
-
-        super().run(legacy_data)
+    def load_blend_data(self):
+        ensure_legacy_blend_data()
