--- FortnitePorting.Plugins/Blender/fortnite_porting/processing/legacy/mappings.py
@@ -1,157 +1,0 @@
-legacy_curve_names = [
-    "C_glabella_down_pose",
-    "tongue_up_pose",
-    "phoneme_mbp_pose",
-    "L_upper_lip_lower_pose",
-    "R_blink_pose",
-    "R_wide_pose",
-    "look_right_pose",
-    "jaw_up_pose",
-    "L_squeeze_pose",
-    "R_brow_up_pose",
-    "look_left_pose",
-    "tongue_back_pose",
-    "R_squint_inner_pose",
-    "look_up_pose",
-    "phoneme_fv_pose",
-    "jaw_back_pose",
-    "R_smile_pose",
-    "R_lip_corner_narrow_pose",
-    "eye_convergence_pose",
-    "R_brow_down_pose",
-    "R_frown_pose",
-    "L_lower_lip_down_pose",
-    "L_blink_pose",
-    "L_lip_corner_narrow_pose",
-    "look_down_pose",
-    "R_upper_lip_raiser_pose",
-    "phoneme_oo_pose",
-    "R_brow_squeeze_in_pose",
-    "jaw_left_pose",
-    "L_lip_corner_wide_pose",
-    "C_nose_wrinkler_pose",
-    "R_lower_lip_up_pose",
-    "jaw_open_pose",
-    "L_brow_up_pose",
-    "R_lip_corner_wide_pose",
-    "L_lower_lip_up_pose",
-    "R_lower_lip_down_pose",
-    "L_upper_lip_raiser_pose",
-    "R_cheek_up_pose",
-    "L_brow_down_pose",
-    "jaw_forward_pose",
-    "L_squint_inner_pose",
-    "tongue_down_pose",
-    "R_upper_lip_lower_pose",
-    "L_brow_squeeze_in_pose",
-    "jaw_right_pose",
-    "mouth_down_pose",
-    "C_glabella_up_pose",
-    "phoneme_ch_pose",
-    "L_frown_pose",
-    "R_squeeze_pose",
-    "tongue_forward_pose",
-    "L_upperLidDown_pose",
-    "R_upperLidDown_pose",
-    "L_wide_pose",
-    "L_smile_pose",
-    "mouth_up_pose",
-    "L_cheek_up_pose",
-    "base_pose"
-]
-
-metahuman_curve_names = [
-    "browDownL",
-    "browDownR",
-    "browLateralL",
-    "browLateralR",
-    "browRaiseInL",
-    "browRaiseInR",
-    "browRaiseOuterL",
-    "browRaiseOuterR",
-    "eyeBlinkL",
-    "eyeBlinkR",
-    "eyeCheekRaiseL",
-    "eyeCheekRaiseR",
-    "eyeLookDownL",
-    "eyeLookDownR",
-    "eyeLookLeftL",
-    "eyeLookLeftR",
-    "eyeLookRightL",
-    "eyeLookRightR",
-    "eyeLookUpL",
-    "eyeLookUpR",
-    "eyeSquintInnerL",
-    "eyeSquintInnerR",
-    "eyeWidenL",
-    "eyeWidenR",
-    "jawBack",
-    "jawChinRaiseDL",
-    "jawChinRaiseDR",
-    "jawChinRaiseUL",
-    "jawChinRaiseUR",
-    "jawFwd",
-    "jawLeft",
-    "jawOpen",
-    "jawOpenExtreme",
-    "jawRight",
-    "lookAtSwitch",
-    "mouthCornerNarrowL",
-    "mouthCornerNarrowR",
-    "mouthCornerPullL",
-    "mouthCornerPullR",
-    "mouthCornerWideL",
-    "mouthCornerWideR",
-    "mouthDown",
-    "mouthFunnelDL",
-    "mouthFunnelDR",
-    "mouthFunnelUL",
-    "mouthFunnelUR",
-    "mouthLipsPurseDL",
-    "mouthLipsPurseDR",
-    "mouthLipsPurseUL",
-    "mouthLipsPurseUR",
-    "mouthLowerLipBiteL",
-    "mouthLowerLipBiteR",
-    "mouthLowerLipDepressL",
-    "mouthLowerLipDepressR",
-    "mouthLowerLipTowardsTeethL",
-    "mouthLowerLipTowardsTeethR",
-    "mouthPressDL",
-    "mouthPressDR",
-    "mouthPressUL",
-    "mouthPressUR",
-    "mouthStretchL",
-    "mouthStretchR",
-    "mouthUp",
-    "mouthUpperLipRaiseL",
-    "mouthUpperLipRaiseR",
-    "mouthUpperLipTowardsTeethL",
-    "mouthUpperLipTowardsTeethR",
-    "noseNasolabialDeepenL",
-    "noseNasolabialDeepenR",
-    "noseWrinkleL",
-    "noseWrinkleR",
-    "tongueDown",
-    "tongueIn",
-    "tongueOut",
-    "tongueUp",
-    "base"
-]
-
-allowed_reorient_children = {
-    "pelvis": ["spine_01"],
-    "spine_01": ["spine_02"],
-    "spine_02": ["spine_03"],
-    "spine_03": ["spine_04"],
-    "spine_04": ["spine_05"],
-    "spine_05": ["neck_01"],
-    "neck_01": ["neck_02"],
-    "neck_02": ["head"],
-    "thigh_r": ["calf_r"],
-    "calf_r": ["foot_r"],
-    "foot_r": ["ball_r"],
-    "thigh_l": ["calf_l"],
-    "calf_l": ["foot_l"],
-    "foot_l": ["ball_l"],
-}
